Rename file name and drop wrong func
